Fixed TensorArray.Write Index Error
Not entirely sure why the index "i+1" is being seen as int64 when it was never casted to be int64 (should be int32 by default).

+++ scripts/Translator.py
@@ -32,7 +32,7 @@ class Translator(tf.Module):
             predicted_id = tf.argmax(predictions, axis = -1) # Grab the highest value in the entire Tensor.
 
             # Add the predicted_id to our output, which was fed to our decoder as input.
-            output_array = output_array.write(i+1, predicted_id[0])
+            output_array = output_array.write(tf.cast(i+1, dtype = tf.int64), predicted_id[0])
 
             # If the predicted token is the end token, then we reached the end of the sentence and should stop generating.
             if predicted_id == end:
